Remove a module with just one function
